ranking

+++ dbApi/routes/devices.js
@@ -5,6 +5,9 @@ const cache = require('../cache');
 const mqtt = require('../broker');
 const TIME_TO_LIVE = process.env.TIME_TO_LIVE || 120;
 
+const ASSISTANT = 1;
+const ENGINEER = 2;
+
 router.get('', (req, res) => {
     try {
         if (req.headers.authorization) {
@@ -14,10 +17,12 @@ router.get('', (req, res) => {
                     return;
                 }
                 if (reply) {
-                    cache.EXPIRE(req.headers.authorization, TIME_TO_LIVE);
-                    let data = await Device.find({}, { _id: 0, __v: 0 });
-                    res.status(200).send(data);
-                    return;
+                    if (reply >= ASSISTANT) {
+                        cache.EXPIRE(req.headers.authorization, TIME_TO_LIVE);
+                        let data = await Device.find({}, { _id: 0, __v: 0 });
+                        res.status(200).send(data);
+                        return;
+                    }
                 }
                 res.sendStatus(401);
                 return;
@@ -41,18 +46,20 @@ router.get('/:tag', (req, res) => {
                     return;
                 }
                 if (reply) {
-                    let device = await Device.findOne(
-                        { tag: req.params.tag },
-                        { _id: 0, __v: 0 }
-                    );
-                    if (device) {
-                        cache.EXPIRE(req.headers.authorization, TIME_TO_LIVE);
-                        res.status(200).send(device);
+                    if (reply >= ASSISTANT) {
+                        let device = await Device.findOne(
+                            { tag: req.params.tag },
+                            { _id: 0, __v: 0 }
+                        );
+                        if (device) {
+                            cache.EXPIRE(req.headers.authorization, TIME_TO_LIVE);
+                            res.status(200).send(device);
+                            return;
+                        } else {
+                            res.sendStatus(204);
+                        }
                         return;
-                    } else {
-                        res.sendStatus(204);
                     }
-                    return;
                 }
                 res.sendStatus(401);
                 return;
@@ -76,27 +83,29 @@ router.post('', (req, res) => {
                     return;
                 }
                 if (reply) {
-                    let body = req.body;
-                    let dupplicated = await Device.findOne(
-                        { $or: [{ serial: body.serial }, { tag: body.tag }] },
-                        {}
-                    );
-                    if (dupplicated) {
-                        res.sendStatus(403);
-                        return;
-                    } else {
-                        let device = new Device({
-                            serial: body.serial,
-                            tag: body.tag,
-                            modbus: body.modbus,
-                            frec: body.frec,
-                            unit: body.unit
-                        });
-                        mqtt.publish(`cmnd/${device.tag}/frec`, `${device.frec}`);
-                        await device.save();
-                        cache.EXPIRE(req.headers.authorization, TIME_TO_LIVE);
-                        res.sendStatus(201);
-                        return;
+                    if (reply >= ENGINEER) {
+                        let body = req.body;
+                        let dupplicated = await Device.findOne(
+                            { $or: [{ serial: body.serial }, { tag: body.tag }] },
+                            {}
+                        );
+                        if (dupplicated) {
+                            res.sendStatus(403);
+                            return;
+                        } else {
+                            let device = new Device({
+                                serial: body.serial,
+                                tag: body.tag,
+                                modbus: body.modbus,
+                                frec: body.frec,
+                                unit: body.unit
+                            });
+                            mqtt.publish(`cmnd/${device.tag}/frec`, `${device.frec}`);
+                            await device.save();
+                            cache.EXPIRE(req.headers.authorization, TIME_TO_LIVE);
+                            res.sendStatus(201);
+                            return;
+                        }
                     }
                 }
                 res.sendStatus(401);
@@ -121,23 +130,25 @@ router.put('/', (req, res) => {
                     return;
                 }
                 if (reply) {
-                    const body = req.body;
-                    const filter = { serial: body.serial };
-                    const update = {
-                        tag: body.tag,
-                        modbus: body.modbus,
-                        frec: body.frec,
-                        unit: body.unit
-                    }
-                    await Device.findOneAndUpdate(filter, update);
-                    let device = await Device.findOne(filter, { _id: 0, __v: 0 });
-                    if (device) {
-                        cache.EXPIRE(req.headers.authorization, TIME_TO_LIVE);
-                        res.status(200).send(device);
-                        return;
-                    } else {
-                        res.sendStatus(403);
-                        return;
+                    if (reply >= ENGINEER) {
+                        const body = req.body;
+                        const filter = { serial: body.serial };
+                        const update = {
+                            tag: body.tag,
+                            modbus: body.modbus,
+                            frec: body.frec,
+                            unit: body.unit
+                        }
+                        await Device.findOneAndUpdate(filter, update);
+                        let device = await Device.findOne(filter, { _id: 0, __v: 0 });
+                        if (device) {
+                            cache.EXPIRE(req.headers.authorization, TIME_TO_LIVE);
+                            res.status(200).send(device);
+                            return;
+                        } else {
+                            res.sendStatus(403);
+                            return;
+                        }
                     }
                 }
                 res.sendStatus(401);
@@ -162,14 +173,16 @@ router.delete('/', (req, res) => {
                     return;
                 }
                 if (reply) {
-                    let deleted = await Device.findOneAndDelete({ tag: req.body.tag }, {});
-                    if (deleted) {
-                        cache.EXPIRE(req.headers.authorization, TIME_TO_LIVE);
-                        res.sendStatus(202);
-                        return;
-                    } else {
-                        res.sendStatus(403);
-                        return;
+                    if (reply >= ENGINEER) {
+                        let deleted = await Device.findOneAndDelete({ tag: req.body.tag }, {});
+                        if (deleted) {
+                            cache.EXPIRE(req.headers.authorization, TIME_TO_LIVE);
+                            res.sendStatus(202);
+                            return;
+                        } else {
+                            res.sendStatus(403);
+                            return;
+                        }
                     }
                 }
                 res.sendStatus(401);
