+++ dbApi/routes/users.js
@@ -7,6 +7,8 @@ const TIME_TO_LIVE = process.env.TIME_TO_LIVE || 120;
 const bcrypt = require('bcrypt');
 const SALT_ROUNDS = process.env.SALT_ROUNDS || 10;
 
+const ADMINISTRATOR = 3;
+
 router.get('/all', (req, res) => {
     try {
         if (req.headers.authorization) {
@@ -17,10 +19,12 @@ router.get('/all', (req, res) => {
                     return
                 }
                 if (reply) {
-                    let data = await User.find({}, { _id: 0, password: 0, __v: 0 });
-                    cache.EXPIRE(req.headers.authorization, TIME_TO_LIVE);
-                    res.status(200).send(data);
-                    return;
+                    if (reply == ADMINISTRATOR) {
+                        let data = await User.find({}, { _id: 0, password: 0, __v: 0 });
+                        cache.EXPIRE(req.headers.authorization, TIME_TO_LIVE);
+                        res.status(200).send(data);
+                        return;
+                    }
                 }
                 res.sendStatus(401);
                 return;
@@ -43,30 +47,33 @@ router.post('/new', (req, res) => {
                     return
                 }
                 if (reply) {
-                    const body = req.body;
-                    let dupplicated = await User.findOne(
-                        { $or: [{ name: body.name }, { email: body.email }] },
-                        {}
-                    );
-                    if (dupplicated) {
-                        res.sendStatus(403)
-                        return;
-                    } else {
-                        const SALT = bcrypt.genSaltSync(SALT_ROUNDS);
-                        const HASH = bcrypt.hashSync(body.password, SALT);
-                        console.log(HASH);
-                        let user = new User({
-                            name: body.name,
-                            email: body.email,
-                            password: HASH,
-                            rank: body.rank
-                        });
-                        await user.save();
-                        cache.EXPIRE(req.headers.authorization, TIME_TO_LIVE);
-                        res.sendStatus(201);
-                        return;
+                    if (reply == ADMINISTRATOR) {
+                        const body = req.body;
+                        let dupplicated = await User.findOne(
+                            { $or: [{ name: body.name }, { email: body.email }] },
+                            {}
+                        );
+                        if (dupplicated) {
+                            res.sendStatus(403)
+                            return;
+                        } else {
+                            const SALT = bcrypt.genSaltSync(SALT_ROUNDS);
+                            const HASH = bcrypt.hashSync(body.password, SALT);
+                            let user = new User({
+                                name: body.name,
+                                email: body.email,
+                                password: HASH,
+                                rank: body.rank
+                            });
+                            await user.save();
+                            cache.EXPIRE(req.headers.authorization, TIME_TO_LIVE);
+                            res.sendStatus(201);
+                            return;
+                        }
                     }
                 }
+                res.sendStatus(401);
+                return;
             });
         } else {
             res.sendStatus(401);
@@ -77,7 +84,7 @@ router.post('/new', (req, res) => {
     }
 });
 
-router.put('/', (req, res) => {
+router.put('/change/rank', (req, res) => {
     try {
         if (req.headers.authorization) {
             cache.GET(req.headers.authorization, async (error, reply) => {
@@ -87,24 +94,104 @@ router.put('/', (req, res) => {
                     return;
                 }
                 if (reply) {
-                    const body = req.body;
-                    let updated = await User.findOneAndUpdate(
-                        { name: body.name },
-                        {
-                            email: body.email,
-                            password: body.password,
-                            rank: body.rank
+                    if (reply == ADMINISTRATOR) {
+                        const body = req.body;
+                        let updated = await User.findOneAndUpdate(
+                            { name: body.name },
+                            { rank: body.rank }
+                        );
+                        if (updated) {
+                            cache.EXPIRE(req.headers.authorization, TIME_TO_LIVE);
+                            res.sendStatus(200);
+                            return;
+                        } else {
+                            res.sendStatus(403);
+                            return;
                         }
-                    );
-                    if (updated) {
-                        cache.EXPIRE(req.headers.authorization, TIME_TO_LIVE);
-                        res.sendStatus(200);
-                        return;
-                    } else {
-                        res.sendStatus(403);
-                        return;
                     }
                 }
+                res.sendStatus(401);
+                return;
+            });
+
+        } else {
+            res.sendStatus(401);
+        }
+    } catch (error) {
+        console.log(err);
+        res.sendStatus(500);
+    }
+});
+
+router.put('/reset/password', (req, res) => {
+    try {
+        if (req.headers.authorization) {
+            cache.GET(req.headers.authorization, async (error, reply) => {
+                if (error) {
+                    console.log(error);
+                    res.sendStatus(401);
+                    return;
+                }
+                if (reply) {
+                    if (reply == ADMINISTRATOR) {
+                        const body = req.body;
+                        const SALT = bcrypt.genSaltSync(SALT_ROUNDS);
+                        const HASH = bcrypt.hashSync("reset", SALT);
+                        let updated = await User.findOneAndUpdate(
+                            { name: body.name },
+                            { password: HASH }
+                        );
+                        if (updated) {
+                            cache.EXPIRE(req.headers.authorization, TIME_TO_LIVE);
+                            res.sendStatus(200);
+                            return;
+                        } else {
+                            res.sendStatus(403);
+                            return;
+                        }
+                    }
+                }
+                res.sendStatus(401);
+                return;
+            });
+
+        } else {
+            res.sendStatus(401);
+        }
+    } catch (error) {
+        console.log(err);
+        res.sendStatus(500);
+    }
+});
+
+router.put('/change/password', (req, res) => {
+    try {
+        if (req.headers.authorization) {
+            cache.GET(req.headers.authorization, async (error, reply) => {
+                if (error) {
+                    console.log(error);
+                    res.sendStatus(401);
+                    return;
+                }
+                if (reply) {
+                    if (reply == ADMINISTRATOR) {
+                        const body = req.body;
+                        let user = await User.findOne({ name: body.name });
+                        if (bcrypt.compareSync(body.oldPassword, user.password)) {
+                            const SALT = bcrypt.genSaltSync(SALT_ROUNDS);
+                            const HASH = bcrypt.hashSync(body.newPassword, SALT);
+                            user.password = HASH;
+                            await user.save({ isNew: false });
+                            res.sendStatus(200);
+                            return;
+                        } else {
+                            res.sendStatus(401);
+                            return;
+                        }
+                    }
+                }
+                res.sendStatus(401);
+                return;
             });
 
         } else {
@@ -126,14 +213,20 @@ router.delete('/', async (req, res) => {
                     return;
                 }
                 if (reply) {
-                    const body = req.body;
-                    let deleted = await User.findOneAndDelete({ name: body.name });
-                    if (deleted) {
-                        cache.EXPIRE(req.headers.authorization, TIME_TO_LIVE);
-                        res.sendStatus(202);
-                    } else {
-                        res.sendStatus(403);
+                    if (reply == ADMINISTRATOR) {
+                        const body = req.body;
+                        let deleted = await User.findOneAndDelete({ name: body.name });
+                        if (deleted) {
+                            cache.EXPIRE(req.headers.authorization, TIME_TO_LIVE);
+                            res.sendStatus(202);
+                            return;
+                        } else {
+                            res.sendStatus(403);
+                            return;
+                        }
                     }
+                    res.sendStatus(401);
+                    return;
                 } else {
                     res.sendStatus(401);
                     return;
